Fixed bug in octospi manager hal init
Without this fix the second instance gets disabled if one of the
IO-Ports is not used. Even thought there will be no conflict between the
instances since the specified port is not used.

Example: Both octospi interfaces are configured to use quad mode.  This
means the DQSPort and IOHighPort will be initialized to be unused for
both instances. In this case the second octospi initialization call will
disable the first instance since it assumes it has the same
configuration which will lead to conflicts.

Same would apply if one decides to use external GPIO chip selects and
leave the NSCPort uninitialized.

To fix this problem, an additional check has been added to first check
if the specific Port is not initialized. In this case it is assumed
there is no conflict for this port and the check proceeds with the next
one.

Additionally the assertion which verifies the NCSPort input has been
changed to accept 0 for the case of an unused NCSPort. This needed an
additional step to verify the passed NCSPort is not 0 before configuring
the PCR register.